pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "chorus-ai"
version = "0.1.0"
description = "Multi-LLM deliberation tool. Have your AI models discuss, debate, and build on each other's ideas — CLI-first, $60/month flat."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "rich>=13.0",
]

[tool.hatch.build.targets.wheel]
packages = ["chorus"]

[project.scripts]
chorus = "chorus.__main__:main"
